backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
python-socketio==5.10.0
aiofiles==23.2.1
uuid==1.30